requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.8
alembic==2.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2

# Optional / Dev
pytest==7.4.2
black==23.12.0
isort==6.0.2
mypy==1.5.1
